# api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BookmarkHive API",
        "description": "Decentralized social bookmarking service based on the Activity Pub protocol.",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://bookmarkhive.test",
            "description": "API Dev Server"
        },
        {
            "url": "https://api2.bookmarkhive.test",
            "description": "API 2 Dev Server"
        }
    ],
    "paths": {
...
        "/users/me/bookmarks": {
            "get": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "List own bookmarks",
                "description": "Returns a paginated collection of bookmarks owned by the authenticated user. Supports filtering by tags and cursor-based pagination. Default page size is 24 items, ordered by creation date (newest first).",
                "operationId": "listOwnBookmarks",
                "parameters": [
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Comma-separated list of tag slugs to filter by",
                        "schema": {
                            "type": "string",
                            "example": "tag-one,tag-two"
                        }
                    },
                    {
                        "name": "after",
                        "in": "query",
                        "description": "Cursor for pagination - bookmark ID to fetch results after",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookmarks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "collection": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookmarkShowPrivate"
                                            }
                                        },
                                        "prevPage": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "nextPage": {
                                            "description": "URL for next page if available",
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "total": {
                                            "type": "integer",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                },
                                "examples": {
                                    "bookmark_list": {
                                        "summary": "Paginated list of bookmarks",
                                        "value": {
                                            "collection": [
                                                {
                                                    "id": "01234567-89ab-cdef-0123-456789abcdef",
                                                    "createdAt": "2024-01-01T12:00:00+00:00",
                                                    "title": "Example Bookmark",
                                                    "url": "https://example.com",
                                                    "domain": "example.com",
                                                    "account": {
                                                        "@iri": "https://bookmarkhive.test/profile/janedoe",
                                                        "username": "janedoe",
                                                        "instance": "bookmarkhive.test",
                                                        "inboxUrl": "https://bookmarkhive.test/profile/janedoe/inbox",
                                                        "outboxUrl": "https://bookmarkhive.test/profile/janedoe/outbox",
                                                        "sharedInboxUrl": "https://bookmarkhive.test/inbox",
                                                        "followerUrl": "https://bookmarkhive.test/profile/janedoe/followers",
                                                        "followingUrl": "https://bookmarkhive.test/profile/janedoe/following"
                                                    },
                                                    "isPublic": true,
                                                    "tags": [
                                                        {
                                                            "name": "Bookmarking",
                                                            "slug": "bookmarking",
                                                            "@iri": "https://bookmarkhive.test/users/me/tags/bookmarking"
                                                        }
                                                    ],
                                                    "instance": "bookmarkhive.test",
                                                    "@iri": "https://bookmarkhive.test/users/me/bookmarks/01234567-89ab-cdef-0123-456789abcdef"
                                                }
                                            ],
                                            "prevPage": null,
                                            "nextPage": "https://bookmarkhive.test/users/me/bookmarks?after=01234567-89ab-cdef-0123-456789abcdef",
                                            "total": null
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Create a new bookmark",
                "description": "Creates a new bookmark. If a bookmark with the same URL already exists, the previous one will be marked as outdated.",
                "operationId": "createBookmark",
                "requestBody": {
                    "description": "Bookmark data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookmarkCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bookmark created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookmarkShowPrivate"
                                },
                                "examples": {
                                    "created_bookmark": {
                                        "summary": "Successfully created bookmark",
                                        "value": {
                                            "id": "01234567-89ab-cdef-0123-456789abcdef",
                                            "createdAt": "2024-01-01T12:00:00+00:00",
                                            "title": "Example Bookmark",
                                            "url": "https://example.com",
                                            "domain": "example.com",
                                            "account": {
                                                "@iri": "https://bookmarkhive.test/profile/janedoe",
                                                "username": "janedoe",
                                                "instance": "bookmarkhive.test",
                                                "inboxUrl": "https://bookmarkhive.test/profile/janedoe/inbox",
                                                "outboxUrl": "https://bookmarkhive.test/profile/janedoe/outbox",
                                                "sharedInboxUrl": "https://bookmarkhive.test/inbox",
                                                "followerUrl": "https://bookmarkhive.test/profile/janedoe/followers",
                                                "followingUrl": "https://bookmarkhive.test/profile/janedoe/following"
                                            },
                                            "isPublic": false,
                                            "tags": [
                                                {
                                                    "name": "Bookmarking",
                                                    "slug": "bookmarking",
                                                    "@iri": "https://bookmarkhive.test/users/me/tags/bookmarking"
                                                }
                                            ],
                                            "instance": "bookmarkhive.test",
                                            "@iri": "https://bookmarkhive.test/users/me/bookmarks/01234567-89ab-cdef-0123-456789abcdef"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "422": {
                        "description": "Validation error - invalid data or invalid tag IRIs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "invalid_tag_iri": {
                                        "summary": "Invalid tag reference",
                                        "value": {
                                            "error": {
                                                "code": 422,
                                                "message": "Invalid tag IRI provided"
                                            }
                                        }
                                    },
                                    "tags_as_objects": {
                                        "summary": "Tags provided as JSON objects instead of IRIs",
                                        "value": {
                                            "error": {
                                                "code": 422,
                                                "message": "Tags must be IRIs, not objects"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/me/bookmarks/{id}": {
            "get": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Get own bookmark by ID",
                "description": "Returns a specific bookmark owned by the authenticated user.",
                "operationId": "getOwnBookmark",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bookmark ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookmark details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookmarkShowPrivate"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "404": {
                        "description": "Bookmark not found or not owned by user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Delete a bookmark",
                "description": "Permanently deletes a bookmark (and all its versions/history) owned by the authenticated user.",
                "operationId": "deleteBookmark",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bookmark ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Bookmark deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "404": {
                        "description": "Bookmark not found or not owned by user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Update a bookmark",
                "description": "Updates an existing bookmark. URL cannot be changed after creation (URL field in request body is ignored). Tags are preserved if not included in the request body.",
                "operationId": "updateBookmark",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bookmark ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Bookmark update data",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookmarkUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bookmark updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookmarkShowPrivate"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "404": {
                        "description": "Bookmark not found or not owned by user"
                    },
                    "422": {
                        "description": "Validation error - invalid data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "invalid_data": {
                                        "summary": "Validation error",
                                        "value": {
                                            "error": {
                                                "code": 422,
                                                "message": "Unprocessable Content"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/users/me/files/{id}": {
            "get": {
                "tags": [
                    "Files"
                ],
                "summary": "Get file object (not supported)",
                "description": "This endpoint is not supported. Use POST to upload files.",
                "operationId": "getFileObject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "File object ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Method not allowed - use POST to upload files"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/me/files": {
            "post": {
                "tags": [
                    "Files"
                ],
                "summary": "Upload a file",
                "description": "Uploads a file and returns a FileObject that can be referenced in bookmarks (mainImage, archive).",
                "operationId": "uploadFile",
                "requestBody": {
                    "description": "File to upload",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "File to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileObject"
                                },
                                "examples": {
                                    "uploaded_file": {
                                        "summary": "Successfully uploaded file",
                                        "value": {
                                            "contentUrl": "https://bookmarkhive.test/storage/files/abc123.jpg",
                                            "size": 102400,
                                            "mime": "image/jpeg",
                                            "@iri": "https://bookmarkhive.test/users/me/files/01234567-89ab-cdef-0123-456789abcdef"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "422": {
                        "description": "Validation error - invalid file",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "invalid_file": {
                                        "summary": "Invalid file validation error",
                                        "value": {
                                            "error": {
                                                "code": 422,
                                                "message": "Unprocessable Content"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BookmarkCreate": {
                "description": "Bookmark creation data",
                "required": [
                    "title",
                    "url"
                ],
                "properties": {
                    "title": {
                        "description": "Bookmark title",
                        "type": "string"
                    },
                    "url": {
                        "description": "Bookmark URL",
                        "type": "string",
                        "format": "uri"
                    },
                    "isPublic": {
                        "description": "Whether the bookmark is public",
                        "type": "boolean",
                        "default": false
                    },
                    "tags": {
                        "description": "Array of tag IRIs (must be valid IRIs pointing to existing tags owned by the user)",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "iri",
                            "example": "https://bookmarkhive.test/users/me/tags/web-development"
                        }
                    },
                    "mainImage": {
                        "description": "IRI of main image FileObject",
                        "type": "string",
                        "format": "iri",
                        "nullable": true
                    },
                    "archive": {
                        "description": "IRI of archive FileObject",
                        "type": "string",
                        "format": "iri",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "BookmarkShowPrivate": {
                "description": "Bookmark object with owner-level details",
                "properties": {
                    "id": {
                        "description": "Bookmark ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "createdAt": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "title": {
                        "description": "Bookmark title",
                        "type": "string"
                    },
                    "url": {
                        "description": "Bookmark URL",
                        "type": "string",
                        "format": "uri"
                    },
                    "domain": {
                        "description": "Extracted domain from URL",
                        "type": "string"
                    },
                    "account": {
                        "$ref": "#/components/schemas/AccountShowPublic"
                    },
                    "isPublic": {
                        "description": "Whether the bookmark is public",
                        "type": "boolean"
                    },
                    "tags": {
                        "description": "Associated tags",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "mainImage": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/FileObject"
                            }
                        ],
                        "nullable": true,
                        "description": "Main image file object",
                        "type": "object"
                    },
                    "archive": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/FileObject"
                            }
                        ],
                        "nullable": true,
                        "description": "Archive file object",
                        "type": "object"
                    },
                    "instance": {
                        "description": "Instance host where the bookmark was created",
                        "type": "string"
                    },
                    "@iri": {
                        "description": "IRI of the bookmark resource",
                        "type": "string",
                        "format": "iri"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token obtained from POST /auth endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}